+++ src/xsdba/adjustment.py
@@ -542,15 +542,23 @@ class EmpiricalQuantileMapping(TrainAdjust):
             standard_name="Model quantiles",
             long_name="Quantiles of model on the reference period",
         )
-        return ds, {"group": group, "kind": kind}
+        if adapt_freq_thresh is None:
+            ds = ds.drop_vars(["P0_ref", "P0_hist", "pth"])
+
+        return ds, {
+            "group": group,
+            "kind": kind,
+            "adapt_freq_thresh": adapt_freq_thresh,
+        }
 
     def _adjust(self, sim, interp="nearest", extrapolation="constant"):
         return qm_adjust(
-            xr.Dataset({"af": self.ds.af, "hist_q": self.ds.hist_q, "sim": sim}),
+            self.ds.assign(sim=sim),
             group=self.group,
             interp=interp,
             extrapolation=extrapolation,
             kind=self.kind,
+            adapt_freq_thresh=self.adapt_freq_thresh,
         ).scen
 
 
@@ -645,6 +653,8 @@ class DetrendedQuantileMapping(TrainAdjust):
             jitter_over_thresh_value=jitter_over_thresh_value,
             jitter_over_thresh_upper_bnd=jitter_over_thresh_upper_bnd,
         )
+        if adapt_freq_thresh is None:
+            ds = ds.drop_vars(["P0_ref", "P0_hist", "pth"])
 
         ds.af.attrs.update(
             standard_name="Adjustment factors",
@@ -658,7 +668,11 @@ class DetrendedQuantileMapping(TrainAdjust):
             standard_name="Scaling factor",
             description="Scaling factor making the mean of hist match the one of hist.",
         )
-        return ds, {"group": group, "kind": kind}
+        return ds, {
+            "group": group,
+            "kind": kind,
+            "adapt_freq_thresh": adapt_freq_thresh,
+        }
 
     def _adjust(
         self,
@@ -674,6 +688,7 @@ class DetrendedQuantileMapping(TrainAdjust):
             detrend=detrend,
             group=self.group,
             kind=self.kind,
+            adapt_freq_thresh=self.adapt_freq_thresh,
         ).scen
         # Detrending needs units.
         scen.attrs["units"] = sim.units
@@ -726,15 +741,17 @@ class QuantileDeltaMapping(EmpiricalQuantileMapping):
 
     def _adjust(self, sim, interp="nearest", extrapolation="constant"):
         out = qdm_adjust(
-            xr.Dataset({"sim": sim, "af": self.ds.af, "hist_q": self.ds.hist_q}),
+            self.ds.assign(sim=sim),
             group=self.group,
             interp=interp,
             extrapolation=extrapolation,
             kind=self.kind,
+            adapt_freq_thresh=self.adapt_freq_thresh,
         )
         if OPTIONS[EXTRA_OUTPUT]:
             out.sim_q.attrs.update(long_name="Group-wise quantiles of `sim`.")
             return out
+
         return out.scen