+++ src/xsdba/base.py
@@ -466,9 +466,10 @@ class Grouper(Parametrizable):
         if isinstance(out, xr.Dataset):
             for name, outvar in out.data_vars.items():
                 if "_group_apply_reshape" in outvar.attrs:
-                    out[name] = self.group(outvar, main_only=True).first(
-                        skipna=False, keep_attrs=True
-                    )
+                    if self.dim in outvar.dims:
+                        out[name] = self.group(outvar, main_only=True).first(
+                            skipna=False, keep_attrs=True
+                        )
                     del out[name].attrs["_group_apply_reshape"]
 
         # Save input parameters as attributes of output DataArray.
@@ -646,12 +647,6 @@ def map_blocks(  # noqa: C901
             for dim in red_dims:
                 reduced_dims.extend(placeholders.get(dim, [dim]))
 
-            for dim in new_dims:
-                if dim in ds.dims and dim not in reduced_dims:
-                    raise ValueError(
-                        f"Dimension {dim} is meant to be added by the "
-                        "computation but it is already on one of the inputs."
-                    )
             if uses_dask(ds):
                 # Use dask if any of the input is dask-backed.
                 chunks = (